Fixing SnmpAttributeTest.
The text was tracking the amount of calls executed against
ipInterfaceDao.load(). I changed it from 3 to 7 (hibernate should take
care about the repetitions). The reason for this is to properly use the
getStorageDir() implementation from DefaultCollectionAgentService inside
of the DefaultCollectionAgent's implementation of the same method, and
cache it like the rest of the variables; instead of having the same
method implemented twice (as it used to be).